Adding excel ext ref correction script
Excel saves spreadsheets/workbooks with absolute external refs, and when
the repo is moved to a new machine, these will be broken. It can
perfectly well understand relative paths, though, so this script opens
up excel workbooks, looks for external references, and updates the paths
so that they are relative to the `cims-models` directory in which this
script lives. If excel files are always stored on the repository with
relative links, then the user should never need to worry about them
(though this script will also be available to the user, and the paths
can be re-set any time).