Use argparse module for better control
We're adding options for dry-run, different levels of failure
strictness, shut-off of logging, and in future there will be subcommands
for *only* checking (not correcting), and for visually outputting
contained external ref links. Also handling of any number of positional
arguments (files/directories) as before. This is much easier with a
proper cmdline argument parsing library.